+++ mcramp/LEGACY/scat/isotropic.py
@@ -49,29 +49,29 @@ class SIsotropic(SPrim):
                                    self.pq_cdf_opencl,
                                    np.uint32(len(self.q)),
                                    np.uint32(len(self.w)),
-                                   np.float64(self.rho),
-                                   np.float64(self.sigma_abs),
-                                   np.float64(self.sigma_scat)).wait()
+                                   np.float32(self.rho),
+                                   np.float32(self.sigma_abs),
+                                   np.float32(self.sigma_scat)).wait()
 
     def _LoadSQW(self, fn):
         with open(fn, 'r') as fin:
             lines = fin.readlines()
-            q = np.fromstring(lines[17], dtype=np.float64, sep=' ')
-            w = np.fromstring(lines[20], dtype=np.float64, sep=' ')
-            rho = np.float64(re.findall(r"[-+]?\d*\.\d+|\d+", lines[6])[0])
-            sigma_abs = np.float64(re.findall(r"[-+]?\d*\.\d+|\d+", lines[9])[0])
-            sigma_coh = np.float64(re.findall(r"[-+]?\d*\.\d+|\d+", lines[10])[0])
+            q = np.fromstring(lines[17], dtype=np.float32, sep=' ')
+            w = np.fromstring(lines[20], dtype=np.float32, sep=' ')
+            rho = np.float32(re.findall(r"[-+]?\d*\.\d+|\d+", lines[6])[0])
+            sigma_abs = np.float32(re.findall(r"[-+]?\d*\.\d+|\d+", lines[9])[0])
+            sigma_coh = np.float32(re.findall(r"[-+]?\d*\.\d+|\d+", lines[10])[0])
 
-        sqw = np.loadtxt(fn, skiprows=23).astype(np.float64) + np.finfo(float).eps.astype(np.float64)
+        sqw = np.loadtxt(fn, skiprows=23).astype(np.float32) + np.finfo(float).eps.astype(np.float32)
         sigma_scat = sigma_coh
 
         pw = simps(sqw.T, axis=1, x=q) / np.linalg.norm(sqw)
-        pw_cdf = np.array([simps(pw[:i], x=w[:i]) for i in range(1,len(w)+1)], dtype=np.float64)
+        pw_cdf = np.array([simps(pw[:i], x=w[:i]) for i in range(1,len(w)+1)], dtype=np.float32)
         pw_cdf = pw_cdf / np.max(pw_cdf)
 
         pq = np.array([sqw[:,i] / np.linalg.norm(sqw[:,i]) for i in range(0,len(w))])
         pq_cdf = np.array(
             [[simps(pq[j,:i], x=q[:i]) / simps(pq[j,:len(q)+1], x=q[:len(q)+1]) \
-            for i in range(1,len(q)+1)] for j in range(0,len(w))], dtype=np.float64)
+            for i in range(1,len(q)+1)] for j in range(0,len(w))], dtype=np.float32)
 
         return (q, w, rho, sigma_abs, sigma_scat, pw_cdf, pq_cdf, sqw)
